Array List implementation of the array sorted polynomial
This implementation sorts the polynomial into an array of set length.  The up side to the implementation is the quick execution of math operations.  The downside is the limited size of the array, and the waste of storage space.  ie.  If you have three elements in the polynomial we are using 1000 storage locatons.